ai-interviewer-backend/app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

SARVAM_API_KEY = os.getenv("SARVAM_API_KEY")
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

# Maximum number of questions in the interview
MAX_INTERVIEW_QUESTIONS = 5